Comment & reply system,Search products

+++ app/Http/Controllers/AdminController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Catagory;
 use App\Models\Product;
 use App\Models\Order;
+use Illuminate\Support\Facades\Auth;
 
 use PDF;
 use Notification;
@@ -15,35 +16,67 @@ class AdminController extends Controller
 {
     public function view_catagory()
     {
-        $data=catagory::all();
-        return view('admin.catagory',compact('data'));
+        if(Auth::id())
+        {$data=catagory::all();
+            return view('admin.catagory',compact('data'));
+        }
+        else
+        {
+            return redirect('login');
+        }
+        
     }
 
     public function add_catagory(Request $request)
     {
+        if(Auth::id())
+        {
         $data=new catagory;
         $data->catagory_name=$request->catagory;
         $data->save();
         return redirect()->back()->with('message','Catagory Added Successfully');
+        }
+        else
+        {
+            return redirect('login');  
+        }
+        
     }
 
     public function delete_catagory($id)
     {
+        if(Auth::id())
+        {
         $data=catagory::find($id);
         $data->delete();
         return redirect()->back()->with('message','Catagory deleted successfully');
+        }
+        else
+        {
+            return redirect('login');  
+        }
+        
     }
 
     public function view_product()
     {
+        if(Auth::id())
+        {
         $catagory= catagory::all();
         return view('admin.product',compact('catagory'));
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function add_product(Request $request)
     {
+        if(Auth::id())
+        {
         $product=new product;
-
         $product->title=$request->title;
         $product->description=$request->description;
         $product->price=$request->price;
@@ -61,33 +94,65 @@ class AdminController extends Controller
         $product->save();
 
         return redirect()->back()->with('message','Product Added Successfully');
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function show_product()
     {
+        if(Auth::id())
+        {
         $product=product::all();
         return view('admin.show_product',compact('product'));
+        }
+        else
+        {
+            return redirect('login');
+        }
+        
     }
 
     public function delete_product($id)
     {
+        if(Auth::id())
+        {
         $product=product::find($id);
 
         $product->delete();
 
         return redirect()->back()->with('message','Product Deleted Successfully');
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function update_product($id)
     {
+        if(Auth::id())
+        {
         $product=product::find($id);
         $catagory=catagory::all();
         return view('admin.update_product',compact('product','catagory'));
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function update_product_confirm(Request $request,$id)
     {
-        $product=product::find($id);
+        if(Auth::id())
+        {
+            $product=product::find($id);
 
         $product->title=$request->title;
         $product->description=$request->description;
@@ -111,39 +176,79 @@ class AdminController extends Controller
         $product->save();
 
         return redirect()->back()->with('message','Product Updated Successfully');
+        }
+        else
+        {
+            return redirect('login');
+        }
+        
     }
 
     public function order()
     {
+        if(Auth::id())
+        {
         $order=order::all();
         return view('admin.order',compact('order'));
+        }
+        else
+        {
+            return redirect('login');
+        }
+        
     }
 
     public function delivered($id)
     {
-        $order=order::find($id);
+        if(Auth::id())
+        {
+            $order=order::find($id);
         $order->delivery_status="delivered";
         $order->payment_status='Paid';
         $order->save();
         return redirect()->back();
+        }
+        else
+        {
+            return redirect('login');
+        }
+        
     }
 
     public function print_pdf($id)
     {
-        $order=order::find($id);
+        if(Auth::id())
+        {
+            $order=order::find($id);
         $pdf=PDF::loadView('admin.pdf',compact('order'));
         return $pdf->download('order_details.pdf');
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function send_email($id)
     {
-        $order=order::find($id);
+        if(Auth::id())
+        {
+            $order=order::find($id);
         return view('admin.email_info',compact('order'));
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function send_user_email(Request $request, $id)
     {
-        $order=order::find($id);
+        if(Auth::id())
+        {
+            $order=order::find($id);
         $details=[
             'greeting'=>$request->greeting,
             'firstline'=>$request->firstline,
@@ -154,14 +259,28 @@ class AdminController extends Controller
         ];
         Notification::send($order,new SendEmailNotification($details));
         return redirect()->back();
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 
     public function searchdata(Request $request)
     {
-        $searchText=$request->search;
+        if(Auth::id())
+        {
+            $searchText=$request->search;
         $order=order::where('name','LIKE',"%$searchText%")->orWhere('phone','LIKE',"%$searchText%")
         ->orWhere('email','LIKE',"%$searchText%")->orWhere('user_id','LIKE',"%$searchText%")
         ->orWhere('product_title','LIKE',"%$searchText%")->get();
         return view('admin.order',compact('order'));
+        }
+        else
+        {
+            return redirect('login'); 
+        }
+        
     }
 }
